requirements.txt
Telethon==1.24.0
python-dotenv==0.20.0